+++ dcso3/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.8
+
+- fix lifetimes on some returned objects were not as long as they should have been
+
 ## 0.0.7
 
 - wherever possible transition error handing to anyhow